+++ tools/pdf-req-parser/main.py
@@ -1,40 +1,31 @@
+from helpers import *
 from PyPDF2 import PdfFileReader
 import simplejson
+import sys
+import os
 
 
-def find_all_entries_between(text, p, q, required_char=None):
-    entries = []
-    entry = ''
-    found = False
-    for ch in text:
-        if ch == p:
-            found = True
-            entry += ch
-        elif ch == q:
-            entry += ch
-            if not required_char:
-                entries.append(entry.replace('{}{}'.format(p, p), p).replace('{}{}'.format(q, q), q))
-            else:
-                if required_char in entry:
-                    entries.append(entry.replace('{}{}'.format(p, p), p).replace('{}{}'.format(q, q), q))
-            entry = ''
-            found = False
-        if found:
-            entry += ch
-    return entries
+if len(sys.argv) < 2:
+    print('Wrong args. Should be: <INPUT_DIRECTORY>')
+    sys.exit(-1)
 
+INPUT_DIRECTORY = sys.argv[1]
+INPUT_LIST_FILE = '{}/input.json'.format(INPUT_DIRECTORY)
+OUTPUT_FILE = '{}/output.json'.format(INPUT_DIRECTORY)
 
-DOCS_DIRECTORY = 'docs'
-INPUT_LIST_FILE = 'input.json'
-OUTPUT_DIRECTORY = 'output'
+if not os.path.isfile(INPUT_LIST_FILE):
+    print('{} file does not exist.'.format(INPUT_LIST_FILE))
+    sys.exit(-2)
 
 input_list = {}
-with open('{}/{}'.format(DOCS_DIRECTORY, INPUT_LIST_FILE), 'rb') as input_file:
+with open('{}'.format(INPUT_LIST_FILE), 'rb') as input_file:
     input_list = simplejson.load(input_file)
 
 requirements = {}
+# Stage 1
+# Processing of the table which is described in input.json file.
 for input_entry in input_list:
-    file_path = '{}/{}'.format(DOCS_DIRECTORY, input_entry['filename'])
+    file_path = '{}'.format(input_entry['filename'])
     print('Processing new file: {}'.format(file_path))
     with open(file_path, 'rb') as f:
         requirements[input_entry['filename']] = {}  # new key in requirements list
@@ -46,7 +37,8 @@ for input_entry in input_list:
             text = page.extractText().replace('\n', '')
             entries.extend(find_all_entries_between(text, '[', ']', required_char='_'))
 
-        # only for SWS files
+        # only for SWS files (temporary)
+        # todo: implement other types (_RS_, _TPS_)
         current_rs = None
         for entry in entries:
             if 'RS' in entry:
@@ -55,5 +47,5 @@ for input_entry in input_list:
             elif 'SWS' in entry:
                 requirements[input_entry['filename']][current_rs][entry] = { 'more_data': 'soon' }
 
-with open('{}/requirements.json'.format(OUTPUT_DIRECTORY), 'w+') as json_file:
+with open(OUTPUT_FILE, 'w+') as json_file:
     json_file.write(simplejson.dumps(requirements, indent=4))
